# server/docs/swagger.yaml
definitions:
...
  responses.Error:
    properties:
      error: {}
    required:
    - error
    type: object
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Timeful.
  title: Timeful API
  version: "1.0"
paths:
...
  /auth/otp/check-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email to check
        in: body
        name: payload
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Checks whether a user with the given email exists
      tags:
      - auth
  /auth/otp/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email to send OTP to
        in: body
        name: payload
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Sends an OTP code to the given email
      tags:
      - auth
  /auth/otp/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email, OTP code, and timezone offset
        in: body
        name: payload
        required: true
        schema:
          properties:
            code:
              type: string
            email:
              type: string
            timezoneOffset:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Verifies an OTP code and signs the user in
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Signs user in and sets the access token session variable
      parameters:
      - description: Object containing the Google authorization code, scope, calendar
          type, and the user's timezone offset
        in: body
        name: payload
        required: true
        schema:
          properties:
            calendarType:
              type: string
            code:
              type: string
            scope:
              type: string
            timezoneOffset:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Signs user in
      tags:
      - auth
  /auth/sign-in-mobile:
    post:
      consumes:
      - application/json
      description: Signs user in and sets the access token session variable
      parameters:
      - description: Object containing the Google authorization code, calendar type,
          and the user's timezone offset
        in: body
        name: payload
        required: true
        schema:
          properties:
            accessToken:
              type: string
            calendarType:
              type: string
            expiresIn:
              type: integer
            idToken:
              type: string
            refreshToken:
              type: string
            scope:
              type: string
            timezoneOffset:
              type: integer
            tokenOrigin:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Signs user in from mobile
      tags:
      - auth
...
  /auth/status:
    get:
      description: Returns a 401 error if user is not signed in, 200 if they are
      responses:
        "200":
          description: OK
        "401":
          description: Error object
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Gets whether the user is signed in or not
      tags:
      - auth
  /auth/visitor-identities:
    post:
      consumes:
      - application/json
      parameters:
      - description: Browser-local public identities; matching HttpOnly credentials
          are required
        in: body
        name: payload
        required: true
        schema:
          properties:
            identities:
              items:
                properties:
                  eventId:
                    type: string
                  eventVisitorId:
                    type: string
                type: object
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      summary: Associate browser Event Visitor Identities with the authenticated account
      tags:
      - auth
...
swagger: "2.0"